@@ -58,30 +58,30 @@
   "dependencies": {
   },
   "devDependencies": {
-    "@typescript-eslint/eslint-plugin": "^6.7.0",
-    "@typescript-eslint/parser": "^6.7.0",
-    "@unocss/eslint-config": "^0.55.7",
-    "eslint": "^8.49.0",
+    "@typescript-eslint/eslint-plugin": "^6.10.0",
+    "@typescript-eslint/parser": "^6.10.0",
+    "@unocss/eslint-config": "^0.57.2",
+    "eslint": "^8.53.0",
     "eslint-config-airbnb": "^19.0.4",
     "eslint-config-airbnb-base": "^15.0.0",
     "eslint-config-airbnb-typescript": "^17.1.0",
-    "eslint-config-next": "^13.4.19",
+    "eslint-config-next": "^14.0.1",
     "eslint-config-prettier": "^9.0.0",
-    "eslint-config-turbo": "^1.10.13",
-    "eslint-define-config": "^1.23.0",
-    "eslint-import-resolver-typescript": "^3.6.0",
-    "eslint-plugin-import": "^2.28.1",
-    "eslint-plugin-jest": "^27.2.3",
-    "eslint-plugin-jsx-a11y": "^6.7.1",
-    "eslint-plugin-prettier": "5.0.0",
+    "eslint-config-turbo": "^1.10.16",
+    "eslint-define-config": "^1.24.1",
+    "eslint-import-resolver-typescript": "^3.6.1",
+    "eslint-plugin-import": "^2.29.0",
+    "eslint-plugin-jest": "^27.6.0",
+    "eslint-plugin-jsx-a11y": "^6.8.0",
+    "eslint-plugin-prettier": "5.0.1",
     "eslint-plugin-react": "7.33.2",
     "eslint-plugin-react-hooks": "^4.6.0",
-    "eslint-plugin-react-refresh": "^0.4.3",
+    "eslint-plugin-react-refresh": "^0.4.4",
     "eslint-plugin-simple-import-sort": "^10.0.0",
     "eslint-plugin-unused-imports": "^3.0.0",
-    "eslint-plugin-vue": "^9.17.0",
+    "eslint-plugin-vue": "^9.18.1",
     "prettier": "^3.0.3",
-    "prettier-plugin-packagejson": "^2.4.5",
-    "vue-eslint-parser": "^9.3.1"
+    "prettier-plugin-packagejson": "^2.4.6",
+    "vue-eslint-parser": "^9.3.2"
   }
 }
