@@ -1,12 +1,6 @@
-module.exports = {
+export default {
   root: true,
-  extends: [
-    'airbnb',
-    'airbnb-typescript',
-    'airbnb/hooks',
-    '@strawberry/eslint-config/base-ts',
-    '@unocss',
-  ],
+  extends: ['airbnb', 'airbnb-typescript', 'airbnb/hooks', '@strawberry/eslint-config/base-ts'],
   plugins: ['react-refresh'],
   rules: {
     'class-methods-use-this': 'off', // 允许类方法不使用this
