@@ -1,5 +1,5 @@
 module.exports = {
   root: true,
   // This tells ESLint to load the config from the package `eslint-config-custom`
-  extends: ['@szmap/eslint-config/strict'],
+  extends: ['@strawberry/eslint-config/strict'],
 };
