@@ -1,18 +1,12 @@
 module.exports = {
   root: true,
-  env: { browser: true, es2020: true },
-  extends: [
-    'eslint:recommended',
-    'plugin:@typescript-eslint/recommended',
-    'plugin:react-hooks/recommended',
-  ],
-  ignorePatterns: ['dist', '.eslintrc.cjs'],
-  parser: '@typescript-eslint/parser',
-  plugins: ['react-refresh'],
+  extends: ['@strawberry/eslint-config/react-app'],
   rules: {
-    'react-refresh/only-export-components': [
-      'warn',
-      { allowConstantExport: true },
-    ],
+    'no-void': 'off',
+    'react/jsx-no-undef': 'off',
+    '@typescript-eslint/no-explicit-any': 'Off',
+  },
+  parserOptions: {
+    project: ['./tsconfig.eslint.json', './tsconfig.json'],
   },
 }
