@@ -1,10 +1,15 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
-import App from './App.tsx'
+
+import { Provider, rootStore } from '@/store'
+
+import App from './App'
 import './index.css'
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
-    <App />
+    <Provider value={rootStore}>
+      <App />
+    </Provider>
   </React.StrictMode>,
 )
